requirements.txt
# requirements.txt
# This file lists all Python packages required by the 'scripts/' 
# and the custom Airflow image build.

# Core libraries for our scripts
requests>=2.31.0      # For making API calls (e.g., CoinGecko)
pandas>=2.1.2         # For data transformation (DataFrames in Silver/Gold)
psycopg2-binary>=2.9.9 # Python library to connect to Postgres

# Libraries for Airflow/Postgres interaction
apache-airflow-providers-postgres>=6.0.0